Dev: add story for TimelineControl and move it to one file
