+++ grapher/controls/Timeline.stories.tsx
@@ -1,0 +1,14 @@
+import * as React from "react"
+
+import { TimelineControl } from "grapher/controls/TimelineControl"
+import { basicGdpGrapher } from "grapher/test/samples"
+
+export default {
+    title: "TimelineControl",
+    component: TimelineControl,
+}
+
+export const Default = () => {
+    const grapher = basicGdpGrapher()
+    return <TimelineControl grapher={grapher} />
+}
