+++ grapher/controls/TimelineControl.scss
