+++ grapher/core/grapher.scss
@@ -134,7 +134,7 @@ $zindex-Tooltip: 20;
 @import "grapher/controls/CommandPalette.scss";
 @import "grapher/controls/ScaleSelector.scss";
 @import "grapher/controls/Controls.scss";
-@import "grapher/controls/Timeline.scss";
+@import "grapher/controls/TimelineControl.scss";
 @import "grapher/controls/ShareMenu.scss";
 @import "grapher/chart/NoDataOverlay.scss";
 @import "grapher/chart/ChartLayout.scss";
